Moving graphql query into own file
